Consolidate API and Auth stacks
There isn't a ton of value in deploying these separately now that
they're tightly coupled. Without this, the logical ID of the UserPool
actually isn't available since it's in another stack.
